+++ src/main/java/com/idt/aio/repository/ProjectAccountRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProjectAccountRepository extends JpaRepository<ProjectAccount, Long> {
+public interface ProjectAccountRepository extends JpaRepository<ProjectAccount, Integer> {
 }
